+++ CHANGES.md
@@ -19,6 +19,7 @@ Changes in 4.4-beta
 - Use ProgramUserConfig() for UPDATE in Grades/Configuration.php, Preferences.php
 - GRADEBOOK_ASSIGNMENTS table: Change DESCRIPTION column type to text in rosariosis.sql & Update.fnc.php
 - GRADEBOOK_ASSIGNMENTS table: Convert DESCRIPTION values from MarkDown to HTML.
+- Change Description field for TinyMCE input in Assignments.php
 
 Changes in 4.3.2
 ----------------